+++ Assets/_Main/_Scripts/Soilders/Bullets/IDamageable.cs
@@ -2,6 +2,6 @@ namespace _Main._Scripts.Soilders.Bullets
 {
     public interface IDamageable 
     {
-         bool TryApplyDamage(int damage);
+         bool TryApplyDamage(float damage);
     }
 }
